nl: reject line number width above INT_MAX via clap value_parser
Per review feedback, bound -w with clap value_parser to match GNU's C int
limit instead of a manual check, so a huge width is rejected at parse time
rather than aborting later with a capacity overflow. Adds a regression test.

Fixes #13347.

+++ src/uu/nl/src/helper.rs
@@ -67,9 +67,11 @@ pub fn parse_options(settings: &mut crate::Settings, opts: &clap::ArgMatches) ->
         Some(Ok(style)) => settings.footer_numbering = style,
         Some(Err(message)) => errs.push(message),
     }
-    match opts.get_one::<usize>(options::NUMBER_WIDTH) {
+    // The upper bound is enforced by clap (see the argument definition); only
+    // the zero case is reported here, to match GNU's message for it.
+    match opts.get_one::<u64>(options::NUMBER_WIDTH) {
         None => {}
-        Some(num) if *num > 0 => settings.number_width = *num,
+        Some(num) if *num > 0 => settings.number_width = *num as usize,
         Some(_) => errs.push(translate!("nl-error-invalid-line-width", "value" => "0")),
     }
     if let Some(num) = opts.get_one::<u64>(options::JOIN_BLANK_LINES) {
